+++ src/two/pre-tagger/model/index.js
@@ -5,6 +5,7 @@ import lexicon from './lexicon/index.js'
 import suffixPatterns from './patterns/suffixes.js'
 import regexNormal from './patterns/regex-normal.js'
 import regexText from './patterns/regex-text.js'
+import regexFallback from './patterns/regex-fallback.js'
 import endsWith from './patterns/endsWith.js'
 import neighbours from './patterns/neighbours.js'
 
@@ -16,6 +17,7 @@ export default {
   suffixPatterns,
   regexNormal,
   regexText,
+  regexFallback,
   endsWith,
   neighbours,
   _multiCache: new Set(),