+++ src/two/pre-tagger/tagger/04-case.js
@@ -10,8 +10,11 @@ const romanNumValid = /^M{0,4}(CM|CD|D?C{0,3})(XC|XL|L?X{0,3})(IX|IV|V?I{0,3})$/
 // if it's a unknown titlecase word, it's a propernoun
 const checkCase = function (term, i) {
   let str = term.text //need case info
-  // titlcase, not first word of sentence
-  if (i !== 0 && titleCase.test(str) === true && hasNumber.test(str) === false && term.tags.has('Date') === false) {
+  // titlecase and not first word of sentence
+  if (i !== 0 && titleCase.test(str) === true && hasNumber.test(str) === false) {
+    term.tags.delete('Verb')
+    term.tags.delete('Adjective')
+    term.tags.delete('Adverb')
     setTag(term, 'ProperNoun', 'titlecase')
     return true
   }