+++ tests/two/tagger/multi.test.js
@@ -12,7 +12,7 @@ test('user-lex-with-hyphenation:', function (t) {
   const sentence =
     'A suicide attack hit the centre of Jardas-al-Abid killing one person (and the attacker) and injuring more than twenty.'
   const found = nlp(sentence, lexicon).match('#Place+')
-  t.equal('jardas al abid', found.eq(0).text('normal'), here + 'found-place1')
+  t.equal('jardas al abid', found.eq(0).text('normal'), here + 'found-hyphen')
   t.equal(lexicon, lexicon, here + 'lexicon-unchanged')
   t.end()
 })
@@ -21,7 +21,7 @@ test('user-lex-with-possessive form:', function (t) {
   const sentence =
     "A suicide attack hit Jardas al Abid's center killing one person (and the attacker) and injuring more than twenty."
   const found = nlp(sentence, lexicon).match('#Place+')
-  t.equal("jardas al abid's", found.eq(0).text('normal'), here + 'found-place2')
+  t.equal("jardas al abid's", found.eq(0).text('normal'), here + 'found-apostrophe')
   t.equal(lexicon, lexicon, here + 'lexicon-unchanged')
   t.end()
 })
@@ -30,7 +30,7 @@ test('user-lex-with-proper name in front:', function (t) {
   const sentence =
     "A suicide attack hit Lybia's Jardas al Abid city killing one person (and the attacker) and injuring more than twenty."
   const found = nlp(sentence, lexicon).match('#Place+')
-  t.equal('jardas al abid', found.eq(0).text('normal'), here + 'found-place3')
+  t.equal('jardas al abid', found.eq(0).text('normal'), here + 'found-proper-name')
   t.equal(lexicon, lexicon, here + 'lexicon-unchanged')
   t.end()
 })
@@ -39,7 +39,7 @@ test('user-lex-with-punctuation:', function (t) {
   const sentence =
     'A suicide attack hit Jardas al Abid, which killed one person (and the attacker) and injured more than twenty.'
   const found = nlp(sentence, lexicon).match('#Place+')
-  t.equal('jardas al abid', found.eq(0).text('normal'), here + 'found-place4')
+  t.equal('jardas al abid', found.eq(0).text('normal'), here + 'found-comma')
   t.equal(lexicon, lexicon, here + 'lexicon-unchanged')
   t.end()
 })